requirements.txt
pip==22.0.2
python-dotenv==1.0.0
setuptools==59.6.0
